refactor: Rename squads package to squad
